requirements.txt
pandas==2.1.4
numpy==1.26.3
yfinance>=0.2.66
ta==0.11.0
groq>=0.4.0
langchain==0.1.6
tiktoken==0.5.2
scikit-learn==1.4.0
matplotlib==3.8.2
seaborn==0.13.1
streamlit==1.31.0
python-dotenv==1.0.1
plotly==5.18.0
rouge-score==0.1.2
jupyter==1.0.0
ipykernel==6.29.0
notebook==7.0.7
httpx==0.27.0

